Added skins, changed to scenebuilder, more to come.
Next time, lets hope for being able to login and launch through the mod
manager. I would hope that hold true. If unable to do that, I want to
redo how forge and stuff works. The less the user has to move the
better, so lets make the forge download from the server. How does that
sound? Good? Good.